[Refactor] 코드 정리
